pyproject.toml
[project]
name = "aristarag"
version = "0.1.0"
description = ""
authors = [
    {name = "Neil Thomas",email = "neilthomas@berkeley.edu"}
]

readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "llama-cloud-services (>=0.6.7,<0.7.0)",
    "llama-parse (>=0.6.4.post1,<0.7.0)",
    "llama-index-core (>=0.12.25,<0.13.0)",
    "llama-index-embeddings-gemini (>=0.3.2,<0.4.0)",
    "llama-index-llms-gemini (>=0.4.12,<0.5.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
